Implement features of int level API.
All the functions of this level accept any kind of registers
in any order. The only limitations that they are treated either
as signed or unsigned, and this has to be decided in advance.

The functions produce refined registers: trimmed, with correctly
set size and signs. Also they provide information about any abnormal
situations that occured during operation (overflow / zero division).